requirements.txt
sqlalchemy
fastapi
strawberry-graphql
strawberry-graphql[fastapi]
uvicorn